K plot
Calculates the estimated weight of fish based on the A and B values from historical length/weight relationships. Calculates K which is the difference between the actual catch weight and the estimated weight for the haul. Calculates the mean and standard deviation of the historical K values. Plots the K values against haul number.